Implement partial attachment.
When restoring windows, we used to block on attaching to the daemon. If the daemon is not feeling well, it can take forever.

Now we wait to restore the windows until we have received child reports from the daemon. We create the job manager as part of this process. Then we can restore the arrangement and marry up the job manager with the PTYTask for the new session.