The previous test for whether windows should be saved was wrong.
NSQuitAlwaysKeepsWindows reflects the setting of the "close windows when quitting an app" system preference. But it can be overridden when restarting or logging out. When you log out, you're asked if you want to restore windows when logging back in. If you turn that on then state should be saved irrsepective of the "close windows when quitting an app" setting.

Luckily there's a friendly private API to access this.

This commit also changes the restoration behavior to be more like macOS: if there is saved stuff, restore it. Previously, we'd delete the saved stuff if NSQuitAlwaysKeepsWindows was off, but obvs that's no good any more. I am amused that the awkward-sounding wording of "close windows when quitting an app" was actually written that way to be a technically correct description of how state saving works.

Issue 9156