Fix background color blending in GPU renderer.
It did not implement the two-layer model correctly. The
model is that there's a view with the default background
color that non-default colors are rendered over top of.

The main change is to perform this blending in the
background color renderer's vertex shader.

This commit also fixes a problem where we'd update the
iTermImageView's content for no reason when switching panes
which made it very slow.

Also fixes a bug where iTermImageView would become visible
when you twiddle the blending slider even though the GPU
renderer was on.

I discovered this because of issue 9245, although it's
probably not what the issue's creator was hoping for.