Don't daemonize iTermServer.
It doesn't seem to be necessary. I tested logging out and back in, fast user switching, and applescript. I was focused on not regressing on issue 9000 (activity monitor shows process not responding; that was because we connected to the window server but didn't run an event loop) and issue 8986 (applescript doesn't work, I think also because of the window server connection).

This should fix issue 9276 where BBEdit wants to look at the parent process of the shell to find the cocoa app that owns it.

The new behavior is similar to what the monoserver did, so I can't think of why it wouldn't work. I look forward to being surprised.