Do synchronous draw when resizing a session. This experiment failed.
The problem I ran in to is that there were really long delays in currentDrawable. The workaround is to do nothing when there are more than 3 drawables outstanding.

Here's the rub: if you decrement the count after the drawable is presented on the metal queue, you get hangs. If you decrement the count from the main thread after presenting, it never hangs. From this I conclude that drawables are only returned to the view in the main thread.

If that is true then trying to use a "metal queue" is a bad idea and I need to move drawing to the main thread.

This will murder performance because drawing is CPU-intensive.

In order to make this work, I need to move token execution off the main thread.