[Python] Remove Splitter.__sessions
In the case of nested splitters, __sessions was not updated
when a child's sessions were changed. I think that
__sessions is an unnecessary optimization that adds a bunch
of complexity including this bug.

It manifested in issue 9784: changing a session's preferred
size had no effect. The reason is that the protobuf was
built by doing a DFS through __children but the tab.sessions
accessor uses __sessions which referred to different
objects.

While this change could affect user code I think the impact
will be limited to scripts affecting nested sessions. I
expect breakage will be on code that worked by accident.