Fix return type of tryToFinishAttachingToMultiserverWithPartialAttachment.
Begin tmux recovery mode in the parser before registering with tasknotifier so input can be handled correctly as soon as it arrives. If the session won't actually be a tmux gateway because we failed to attach to a runnign server, cancel tmux recovery mode right away.

Pause the token executor before enabling the terminal so initial setup can complete before the first token is executed.

Fix a bug in how VT100TmuxParser ignored its first line of input in recovery mode.