Make iTermTuple nullability-correct. This forces fixing up SetUserVariableTrigger.swift.
Make side-effect state into a bitmask instead of a dictionary because it was wasting 3.5% of cpu on the spam.cc benchmark.

Add assertions that out-of-order side effects never occur (i.e., that addSideEffect in joined threads only happens when no side-effects are queued up).

Fix a bug in swapOnscreenIntervalTreeObjects where its `temp` tree could get deallocated causing its doppelgangers not to be removed triggering an assertion when later adding them to the saved tree.