Fix two bugs with minimum contrast: (issue 10273)
1. The prefs panel showed the wrong value when not using separate light/dark
colors because the "use different colors for light and dark mode" checkbox was
not initialized before it was used, causing it to behave as though it was
always checked when looking up values to initialize the min contrast slider.

2. Colormap mutations before the mutable state's side-effect performing
delegate was assigned didn't update the main-thread copy. Fix this by running
those side effects even when there is no delegate since it wasn't used anyway.