Composer improvements.
Rename "Convert Matches to Selections" to "Select
Matches", which is the standard. Use
performFindPanelAction: instead of
convertMatchesToSelections: so Select Matches will
work in the composer as well as in PTYTextView.
Remove performFindPanelAction: from PseudoTerminal
since it really belongs in PTYTextView. Add a
disgusting hack to iTermComposerView to make
Select Matches work in the composer.

ARCify iTermWarning so I can use it from Swift.
Add mullability annotations to it.

Add CommandExplainer to centrailze explainshell
URL logic.

Make cmd-click on a command in the composer open
explainshell for just that command. The logic is
now mostly in ComposerTextView, but it uses a URL
with the iterm2 scheme so some of it ended up in
the app delegate as well.

Change composer's help to describe shortcuts.

Add iTermTextPopoverViewController.sizeToFit and
beat AppKit's buggy sizing code into submission.