Set deployment target for libssh2 correctly.
Bump to xcode 14.3

Fix new warnings.

Fix bug in it_fromKeyValueCodedData which always failed if there were non-dictionary values. Not sure if this is a change in the 14.3 SDK or if it never worked.

Fix boolean instead of logical OR in PTYMouseHandler.

Fix rebuild-deps-if-needed to only update last-xcode-version if the build succeeds.