Fix checking python version in ssh. For some reason running the command as $* gives strange output on my Linux host.
Simplify how CPU utilization is distributed. There was a bug where routers ended up with too many consumers. I never really got to the bottom of it because it was just architected badly.

iTermLocalCPUUtilizationPublisher is now responsible for publishing local CPU usage. It goes to each iTermCPUUtilization which distributes it to routers and then to consumers. SSH-based values from from the SSH process info provider, which takes a similar role to iTermLocalCPUUtilizationPublisher.

When the underlying publisher changes, the status bar component restarts its drawing state.

Also avoid drawing when there is no window.

Issue 11216