Fix unrecognized OSC 9 ; 4 sequences
Recently, Ghostty and many terminals implemented the progress bar status
reporting using OSC 9 ; 4 ; n ; m sequences, where n is a number indicating
the progress type and m is an optional value from 0 to 100 indicating the
progress percentage.

This feature originated from ConEmu[^1] and has been adopted by other
terminal emulators. However, it conflicts with OSC 9 notification
sequence, and that causes issues in iTerm2 where it treats the entire
sequence as a notification leading to spammy notifications with weird
messages.

This commit makes iTerm2 ignore parameterized OSC 9 sequences since they
are not supported. If the sequence is just OSC 9 without parameters, or
if the parameter is not a number, it will be treated as a notification
as before.

[^1]: https://conemu.github.io/en/AnsiEscapeCodes.html#ConEmu_specific_OSC